# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Chatquest",
    "description": "The chatquest API documentation"
  },
  "basePath": "/api/",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "ApiKey": {
      "type": "apiKey",
      "name": "x-api-key",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [],
      "ApiKey": []
    }
  ],
...
  "schemes": ["http", "https"],
  "paths": {
...
    "/messages/{id}": {
      "get": {
        "tags": ["messages"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "minimum": 1,
            "description": "Message ID"
          }
        ],
        "description": "Get a message",
        "operationId": "message",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Response": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "uuid": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "is_temporary": {
          "type": "boolean",
          "default": false
        },
        "friends": {
          "type": "array",
          "default": [],
          "items": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "Message": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "room": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "content": {
          "type": "string"
        },
        "is_edited": {
          "type": "boolean",
          "default": false
        },
        "is_vocal": {
          "type": "boolean",
          "default": false
        },
        "is_image": {
          "type": "boolean",
          "default": false
        },
        "from_server": {
          "type": "boolean",
          "default": false
        },
        "reactions": {
          "type": "array",
          "default": [],
          "items": {
            "$ref": "#/definitions/Reaction"
          }
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Reaction": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "content": {
          "type": "string",
          "default": "👍"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
  }
}